+++ spec/dummy/app/models/user.rb
@@ -1,4 +1,11 @@
 # frozen_string_literal: true
 
+ActiveRecord::Schema.define do
+  create_table :users, force: true do |t|
+    t.string :name
+    t.string :secret
+  end
+end
+
 class User < ActiveRecord::Base
 end