+++ spec/support/with_grpc_server.rb
@@ -4,7 +4,8 @@ shared_context "with gRPC server", :with_grpc_server do
   before(:all) do
     @server = AnyCable::Server.new(
       host: AnyCable.config.rpc_host,
-      **AnyCable.config.to_grpc_params
+      **AnyCable.config.to_grpc_params,
+      interceptors: AnyCable.middleware.to_a
     )
 
     @server.start